flashbanged implemented
flashbang added
