+++ Servo.py
@@ -1,6 +1,6 @@
 import RPi.GPIO as GPIO
-from playwright.sync_api import Playwright, sync_playwright
 import time
+import requests
 
 #Moves the servo back and forth 10 times
 def feeding_time(freq):
@@ -21,14 +21,7 @@ def feeding_time(freq):
         p.stop()
         GPIO.cleanup()
 
-def flash_bang():
-    with sync_playwright() as playwright:
-        browser = playwright.chromium.launch(headless=True)
-        page = browser.new_context().new_page()
-
-        page.goto("http://10.2.177.147:8080/")
-
-        button = page.query_selector("#flashbtn")
-        button.click()
-
-        browser.close()
+def flashbang():
+    url = 'http://localhost:5000/flashbang'
+    x = requests.get(url)
+    return(x.json())
