+++ flask_server/.env
@@ -1,0 +1,2 @@
+FLASK_APP=project/__init__.py
+FLASK_DEBUG=1
