+++ flask_server/services/fishgo_service/Pipfile
@@ -1,0 +1,15 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+flask = "~=2.2.2"
+gunicorn = "~=20.1.0"
+playwright = "~=1.30.0"
+pycryptodome = "~=3.15.0"
+
+[dev-packages]
+
+[requires]
+python_version = "3.10.7"
