+++ flask_server/services/fishgo_service/dockerfile
@@ -1,0 +1,24 @@
+# pull official base image
+FROM python:3.10.7-slim-buster
+
+# set work directory
+WORKDIR /usr/src/app
+
+# set environment variables
+ENV PYTHONDONTWRITEBYTECODE 1
+ENV PYTHONUNBUFFERED 1
+
+# install system dependencies
+RUN apt-get update && apt-get -y install libnss3 libatk-bridge2.0-0 libdrm-dev libxkbcommon-dev libgbm-dev libasound-dev libatspi2.0-0 libxshmfence-dev
+RUN apt-get update && apt-get -y install libcups2 libxcomposite1 libxdamage1 libxfixes3 libxrandr2 libpango-1.0-0 libcairo2
+RUN apt-get update && apt-get install -y netcat
+
+# install dependencies
+RUN pip install --upgrade pip pipenv 
+COPY Pipfile* ./
+
+RUN pipenv install --system --ignore-pipfile
+
+# copy projectd
+COPY . /usr/src/app/
+
