+++ flask_server/services/fishgo_service/manage.py
@@ -1,0 +1,8 @@
+from flask.cli import FlaskGroup
+
+from project import app
+
+cli = FlaskGroup(app)
+
+if __name__ == "__main__":
+    cli()
