+++ flask_server/services/fishgo_service/project/config.py
@@ -1,0 +1,8 @@
+import os
+
+
+basedir = os.path.abspath(os.path.dirname(__file__))
+
+
+class Config(object):
+    pass
